+++ TODO.md
@@ -1,8 +1,8 @@
 # TODO
 - [ ] /web -> /template
-- [ ] /internal/handlers -> /controller (index.go, course.go, dashboard.go)
-- [ ] /internal/model -> /model (user.go, course.go, assignment.go (All DB-related))
-- [ ] /cmd/router/main.go -> main.go
-- [ ] Routes in main.go -> /route/route.go@routes()
+- [x] /internal/handlers -> /controller (index.go, course.go, dashboard.go)
+- [x] /internal/model -> /model (user.go, course.go, assignment.go (All DB-related))
+- [x] /cmd/router/main.go -> main.go
+- [x] Routes in main.go -> /route/route.go@routes()
 - [ ] /db -> /shared/database (or /shared/db, your choice)
 - [x] /internal/util/session.go -> /shared/session/session.go
