+++ db/authdb.go
@@ -2,7 +2,7 @@ package db
 
 import (
 	"fmt"
-	"github.com/JohanAanesen/NTNU-Bachelor-Management-System-For-CS-Assignments/internal/model"
+	"github.com/JohanAanesen/NTNU-Bachelor-Management-System-For-CS-Assignments/model"
 	"golang.org/x/crypto/bcrypt"
 	"log"
 )
